requirements/prod.txt
aiogram==3.17.0
asyncpg==0.30.0 
python-dotenv==1.0.1 
SQLAlchemy==2.0.40